Phase F4 — Rename *-v2.ts files to canonical names
The legacy `mechanical.ts` / `arcane.ts` / `greenward.ts` / `snake-eyes.ts`
data files were deleted in F3. The aspect-shape counterparts shed
their `-v2` suffix and reclaim the canonical name:

  src/data/campaigns/mechanical-v2.ts → mechanical.ts
  src/data/campaigns/arcane-v2.ts     → arcane.ts
  src/data/campaigns/greenward-v2.ts  → greenward.ts
  src/data/campaigns/snake-eyes-v2.ts → snake-eyes.ts

Import-site fixes across 10 files (main.ts side-effects, the two
map test files, the aspect runtimes that import campaign types,
the four runtime test files, WorldMutator's doc comments). No
code changes — pure rename + import-path swap.

Validation:
  • npx tsc --noEmit clean
  • vitest full suite: 1270/1270 pass
  • playwright m10-overthrow + greenward-smoke: 8/8 pass on
    desktop + mobile

Per docs/campaign-aspects-refactor-prd.md Phase F, sub-step F4.
