+++ src/data/campaigns/greenward.ts
