+++ src/data/campaigns/mechanical.ts
